Fixed errors

+++ realestate/realestateapp/views.py
@@ -12,7 +12,10 @@ from django.views.decorators.debug import sensitive_post_parameters
 # Create your views here.
 @csrf_exempt
 def coupons(request):
+    """Ability to view, delete, or create coupons."""
     if request.method == 'GET':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         coupons = Coupon.objects.all()
         couponsList = []
         for coupon in coupons:
@@ -26,6 +29,8 @@ def coupons(request):
             couponsList.append(couponTemp)
         return JsonResponse(couponsList, safe=False, status=200)
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         try:
             data = json.loads(request.body.decode('utf-8'))
         except json.JSONDecodeError:
@@ -36,6 +41,7 @@ def coupons(request):
                 couponT.couponTypeName = data['couponTypeName']
                 couponT.couponTypeDescription = data['couponTypeDescription']
                 couponT.save()
+            #If coupon type has already been made, then retrieve it.
             except:
                 couponT = CouponType.objects.get(couponTypeName=data['couponTypeName'])
             newCoupon = Coupon()
@@ -53,14 +59,21 @@ def coupons(request):
             }
             return JsonResponse(couponJSON, status=200)
     elif request.method == 'DELETE':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         coupons = Coupon.objects.all()
+        #Delete all of the coupons.
         for coupon in coupons:
             coupon.delete()
         return HttpResponse("All coupons were deleted.", status=200)
 
 @csrf_exempt
 def properties(request):
+    """Create a property listing, view all of your own properties, or delete all of your properties"""
     if request.method == 'GET':
+        #Get properties user owns
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         ownProperties = Ownership.objects.filter(user_id=request.user)
         propertiesList = []
         for ownership in ownProperties:
@@ -77,8 +90,10 @@ def properties(request):
                 "propertyBathrooms":prop.propertyBathrooms
             }
             propertiesList.append(propertyTemp)
-        return JsonResponse(propertiesList, status=200)
+        return JsonResponse(propertiesList, safe=False, status=200)
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         try:
             data = json.loads(request.body.decode('utf-8'))
         except json.JSONDecodeError:
@@ -89,6 +104,7 @@ def properties(request):
                 propertyT.propertyTypeName = data['propertyTypeName']
                 propertyT.propertyTypeDescription = data['propertyTypeDescription']
                 propertyT.save()
+            #If property type already exists, then retrieve it.
             except:
                 propertyT = PropertyType.objects.get(propertyTypeName=data['propertyTypeName'])
             try:
@@ -96,6 +112,7 @@ def properties(request):
                 neighborhood.neighborhood_name = data['neighborhoodName']
                 neighborhood.neighborhood_desc = data['neighborhoodDescription']
                 neighborhood.save()
+            #If neighborhood already exists, then retrieve it.
             except:
                 neighborhood = neighborhood.objects.get(neighborhood_name=data['neighborhoodName'])
             propTemp = Property()
@@ -128,7 +145,17 @@ def properties(request):
             propPhoto.property_id = propTemp
             propPhoto.save()
 
+            owner = Ownership()
+            owner.user_id = request.user
+            owner.property_id = propTemp
+            owner.ownershipAskingPrice = data['ownershipAskingPrice']
+            owner.ownershipPaidPrice = data['ownershipPaidPrice']
+            owner.save()
+
             propertyJSON = {
+                "owner":owner.user_id.username,
+                "ownerAskingPrice":owner.ownershipAskingPrice,
+                "ownerPaidPrice":owner.ownershipPaidPrice,
                 "propertyTypeName":propTemp.propertyType.propertyTypeName,
                 "propertyTypeDescription":propTemp.propertyType.propertyTypeDescription,
                 "neighborhoodName":propTemp.neighborhood.neighborhood_name,
@@ -145,14 +172,21 @@ def properties(request):
             }
             return JsonResponse(propertyJSON, status=200)
     elif request.method == 'DELETE':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         ownProperties = Ownership.objects.filter(user_id=request.user)
+        #Delete only properties that user owns
         for ownership in ownProperties:
-            ownership.property_id.delete()
+            prop = Property.objects.get(id=ownership.property_id.id)
+            prop.delete()
         return HttpResponse("All of your properties were deleted.", status=200)
 
 @csrf_exempt
 def specificProperty(request, property_id):
+    """Add property to cart, get property information, or delete specified property"""
     if request.method == 'GET':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         propTemp = Property.objects.get(id=property_id)
         propertyJSON = {
             "propertyTypeName":propTemp.propertyType.propertyTypeName,
@@ -169,6 +203,8 @@ def specificProperty(request, property_id):
         }
         return JsonResponse(propertyJSON, status=200)
     elif request.method == 'POST':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         cart = Cart()
         cart.user_id = request.user
         prop = Property.objects.get(id=property_id)
@@ -176,6 +212,8 @@ def specificProperty(request, property_id):
         cart.save()
         return HttpResponse("Successfully added property to your cart", status=200)
     elif request.method == 'DELETE':
+        if not request.user.is_authenticated:
+            return HttpResponse("Not logged in.", status=401)
         try:
             prop = Property.objects.get(id=property_id)
         except:
